fix : router 구조 변경
- 개별 레이아웃 적용되도록 변경
