+++ src/App.tsx
@@ -1,14 +1,13 @@
-import { Outlet } from 'react-router-dom';
-import MainHeader from './components/Header/MainHeader/MainHeader';
-import SearchBar from './components/SearchBar/SearchBar';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { RouterProvider } from 'react-router-dom';
+import router from './routes/router';
 
+const queryClient = new QueryClient();
 const App = () => {
   return (
-    <div className="flex flex-col items-center web:w-[1440px] web:mx-auto mobile:max-w-full mobile:min-w-[480px]">
-      <MainHeader />
-      <SearchBar />
-      <Outlet />
-    </div>
+    <QueryClientProvider client={queryClient}>
+      <RouterProvider router={router} />
+    </QueryClientProvider>
   );
 };
